+++ app/_layout.js
@@ -1,12 +1,7 @@
 import { View, Text, Button } from 'react-native';
 import React from 'react';
-import {
-  NavigationContainer,
-  Stack,
-  Screen,
-  useRouter,
-  Slot,
-} from 'expo-router';
+import { useRouter } from 'expo-router';
+import { Stack } from 'expo-router/stack';
 
 export default function Layout() {
   const router = useRouter();
@@ -63,7 +58,18 @@ export default function Layout() {
           ),
         }}
       />
+      <Stack.Screen
+        name="evaluationRoll"
+        options={{
+          headerTitle: 'RocknRoll',
+          headerShown: true,
+        }}
+      />
       <Stack.Screen name="tabs" options={{ headerShown: false }} />
+      <Stack.Screen
+        name="evaluationComponents"
+        options={{ headerShown: true }}
+      />
     </Stack>
   );
 }
