+++ app/evaluationComponents/_layout.js
@@ -23,9 +23,6 @@ const EvaluationLayout = () => {
         },
       }}
     >
-      {/* // Stackscreens können so umbenannt werden oder ausgeblendet mit
-      //headerShown */}
-
       <Stack.Screen
         name="evaluationYaw"
         options={{
